+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/AttachmentStoringProcessor.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,9 +21,12 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  */
 
 import com.google.common.collect.ImmutableSet;
+import com.walmartlabs.concord.server.org.project.ProjectDao;
+import com.walmartlabs.concord.server.org.project.ProjectEntry;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
 
+import javax.inject.Inject;
 import javax.inject.Named;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -31,6 +34,8 @@ import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
+import java.util.regex.Pattern;
 
 @Named
 public class AttachmentStoringProcessor implements PayloadProcessor {
@@ -38,8 +43,17 @@ public class AttachmentStoringProcessor implements PayloadProcessor {
     /**
      * Attachments used by the server. They should not be added to the workspace.
      */
-    public static final Set<String> SYSTEM_ATTACHMENT_NAMES = ImmutableSet.of(Payload.WORKSPACE_ARCHIVE.name(),
+    private static final Set<String> SYSTEM_ATTACHMENT_NAMES = ImmutableSet.of(Payload.WORKSPACE_ARCHIVE.name(),
             RequestDataMergingProcessor.REQUEST_ATTACHMENT_KEY.name());
+
+    private static final Pattern RAW_PAYLOAD_PATTERN = Pattern.compile("^(concord|flows)[/.](.*)(yml|yaml)$");
+
+    private final ProjectDao projectDao;
+
+    @Inject
+    public AttachmentStoringProcessor(ProjectDao projectDao) {
+        this.projectDao = projectDao;
+    }
 
     @Override
     public Payload process(Chain chain, Payload payload) {
@@ -55,6 +69,11 @@ public class AttachmentStoringProcessor implements PayloadProcessor {
                 continue;
             }
 
+            UUID projectId = payload.getHeader(Payload.PROJECT_ID);
+            if (hasRawPayload(name) && !canAcceptRawPayload(projectId)) {
+                throw new ProcessException(payload.getProcessKey(), "Project is not accepting flows in attachments (" + name + "). Check the \"Allow payload archives\" setting.");
+            }
+
             Path src = entry.getValue();
             Path dst = workspace.resolve(name);
 
@@ -68,5 +87,18 @@ public class AttachmentStoringProcessor implements PayloadProcessor {
         }
 
         return chain.process(payload);
+    }
+
+    private boolean hasRawPayload(String name) {
+        return RAW_PAYLOAD_PATTERN.matcher(name).matches();
+    }
+
+    private boolean canAcceptRawPayload(UUID projectId) {
+        if (projectId == null) {
+            return true;
+        }
+
+        ProjectEntry project = projectDao.get(projectId);
+        return project.getAcceptsRawPayload();
     }
 }